pyproject.toml
[tool.poetry]
name = "muquiz-boy"
version = "0.1.0"
description = ""
authors = ["Bietola <dincio.montesi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.4.0"
python-telegram-bot = "^13.11"
midi2audio = "^0.1.1"
pydub = "^0.25.1"
abjad = "^3.6"
lenses = "^1.1.0"
dotty-dict = "^1.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
muquiz-boy = 'muquiz_boy.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
